refactor: Extract Go to Anything search logic into a new `useSearch` hook.
